oasis2/requirements.txt
aiohttp==3.7.2
sqlalchemy==1.4.0b1
aioredis==1.3.1
asyncssh==2.4.2
Pygments==2.8.0 
jellyfish==0.8.2 
mypy-extensions==0.4.3
prettytable==2.0.0 
prompt-toolkit==3.0.16
pyparsing==2.4.7 
python-nubia==0.2b5
termcolor==1.1.0 
typing-inspect==0.6.0
wcwidth==0.2.5
pymysql==1.0.2
ks3sdk==1.0.19